Extract widget-preview into its own module
