--- main/src/main/kotlin/com/tasomaniac/devwidget/main/WidgetListData.kt
@@ -1,6 +1,0 @@
-package com.tasomaniac.devwidget.main
-
-import com.tasomaniac.devwidget.data.Widget
-import com.tasomaniac.devwidget.widget.DisplayApplicationInfo
-
-internal data class WidgetListData(val widget: Widget, val apps: List<DisplayApplicationInfo>)
