+++ settings.gradle
@@ -5,5 +5,6 @@ include ':settings'
 include ':base'
 include ':data'
 include ':widget'
+include ':widget-preview'
 
 include ':test-support'
