+++ widget-preview/build.gradle
@@ -1,0 +1,10 @@
+apply from: configFile('android-library-config.gradle')
+
+android {
+    resourcePrefix 'widget_preview'
+}
+
+dependencies {
+    implementation project(':widget')
+    implementation project(':data')
+}
