+++ widget-preview/src/main/AndroidManifest.xml
@@ -1,0 +1,7 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.tasomaniac.devwidget.widget.preview">
+
+    <application
+        android:supportsRtl="true"
+        android:theme="@style/Theme.DevWidget" />
+</manifest>
