+++ widget-preview/src/main/kotlin/com/tasomaniac/devwidget/widget/preview/AppViewHolder.kt
@@ -1,4 +1,4 @@
-package com.tasomaniac.devwidget.main
+package com.tasomaniac.devwidget.widget.preview
 
 import android.view.View
 import android.view.ViewGroup
