+++ widget-preview/src/main/kotlin/com/tasomaniac/devwidget/widget/preview/WidgetListData.kt
@@ -1,0 +1,6 @@
+package com.tasomaniac.devwidget.widget.preview
+
+import com.tasomaniac.devwidget.data.Widget
+import com.tasomaniac.devwidget.widget.DisplayApplicationInfo
+
+data class WidgetListData(val widget: Widget, val apps: List<DisplayApplicationInfo>)
