+++ widget-preview/src/main/res/values/widget_preview_dimens.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="widget_preview_min_height">128dp</dimen>
+</resources>
